One floating-point type: the judgement pass, and what keeps its double
Finishes the conversion. The GUI, the crossfades, the upscalers, the fonts,
the textures and the audio helpers come over, and the three groups that a
float provably cannot carry stay as they are:

  - the wall clock and the frame timings, where the value grows for the whole
    session and what is read off it is the difference of two readings;
  - the scrolling offsets the weather and the title clouds carry, which grow
    with the level's clock - wrapTextureOffset stays double and every caller
    takes a float back;
  - Mat4's perspective, lookAt and in-place operations, which reproduce GLU's
    own arithmetic so a matrix is the numbers GL built.

The renderer's corner baking stays float, measured rather than assumed:
against a sum promoted to double and rounded once, the two differ for 41% of
corners but by at most 0.000488 px, so 0.89% survive the 1/256 subpixel grid
the rasterizer snaps to and no oracle scene moves a pixel. It is four calls a
quad in the renderer's hottest arithmetic.

Nine of the nineteen oracle scenes move, none by more than 8 channel levels of
255 and all of it speckle at one level. Three causes: pointJitter, the
fract(sin(x) * large) hash the beam shines take their per-point size from,
resolves 1/1024 in float and so draws different but equally uniform values -
rms 0.5772 against the ideal 0.5774 over 200 points of each of 2001 seeds,
deciles flat within 3%, neighbours correlated at -0.002 - which is night,
select and the cube that captures them; the credits' stars round to a
different subpixel; and a literal reading 0.1f where it read 0.1 moves a phase
by a part in 10^8, which is the lava, the star wipe, the toxic grid, the CRT
filter and the options dialog, 8 to 40 pixels each.

verify.py and the mingw gate are clean over all 122 sources, both builds link,
and the native and browser smoke tests pass. ROADMAP 57 is done; rendering.md,
upscalers.md and FINDINGS.md follow the types they name.

+++ Blocks5/src/audiocapture.cpp
@@ -410,7 +410,7 @@ void AudioCaptureImpl::convertAndPush(const BYTE* p_data, int numFrames, bool si
 		// through unchanged.
 		while(resamplePos < 1.0f)
 		{
-			const float t = (float)resamplePos;
+			const float t = resamplePos;
 			scratch[2 * numInScratch    ] = floatToShort(prevLeft  + (left  - prevLeft ) * t);
 			scratch[2 * numInScratch + 1] = floatToShort(prevRight + (right - prevRight) * t);
 			numInScratch++;
