+++ Blocks5/src/cf_blend.cpp
@@ -14,5 +14,5 @@ void CF_Blend::render(float t,
 					  uint newImageID)
 {
 	// the old image, fading out over the new one already on the screen
-	drawImage(oldImageID, Vec4f(1.0f, 1.0f, 1.0f, static_cast<float>(1.0f - t)));
+	drawImage(oldImageID, Vec4f(1.0f, 1.0f, 1.0f, 1.0f - t));
 }
